ChangeLog: corrections and minor cleanup of 23.10 release
When released the YANG status section claimed IPv4/IPv6 MTU was
not-supported, which was very confusing since the following Changes
section clearly states added support for it.

The YANG Status section was wrong.  It has also been cleaned up a bit.
E.g., no need to mention which native model provides an augment or a
deviation, to an end-user the only thing that matters is what is added
to or removed from the standard models.

[skip ci] becuase none of this warrants a rebuild.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>