Wrap Buildroot 'legal-info' build target to include Linux license
When building a Linux kernel that does not match the latest kernel known
by Buildroot, the license file is disabled and nothing is extracted.
For an introduction to this see Buildroot gaf33b1c.

To work around this in Infix we wrap the 'legal-info' build target to
set the license file variable and trigger the license extraction anyway.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>